chore: Move helper functions to separate files
